refactor(protocol): move IPC line framer from core to DanTermProtocol
Phase 2 of the pure-core/portable-support refactor. Moves the pure IpcLineFramer +
IpcFrameEvent types out of DanTermCore's IpcConnection.swift into a new
DanTermProtocol source, promoting exactly init()/append()/maxLineBytes/the enum to
public (the only public surface the split adds). The IpcConnection socket class stays
in core for now (Phase 3 moves it to DanTermSupport) and resolves the framer via its
existing `import DanTermProtocol`. Relocates the framer-only test suite to
DanTermProtocolTests as XCTest, matching that package's convention.

No app changes, no manifest changes. CLI/IPC surface unchanged, so no SKILL.md edit.
just test + just build green.
